Default disabled Phytium drivers on OpenCloudOS 8.8
This commit documents the Phytium kernel driver module that is
disabled by default on OpenCloudOS 8.8. Users can configure and
enable these modules as needed.

The reason why the module is disabled by default: It may cause conflicts
on platforms other than Phytium, or it is disabled by default due
to single and rare usage scenarios, in order to improve the efficiency
of kernel building.

Signed-off-by: Jiakun Shuai <shuaijiakun1288@phytium.com.cn>